Update DTO classes to use strict types and camelCase property names

+++ src/HeaderSelector.php
@@ -39,8 +39,8 @@ class HeaderSelector
 {
     /**
      * @param string[] $accept
-     * @param string   $contentType
-     * @param bool     $isMultipart
+     * @param string $contentType
+     * @param bool $isMultipart
      * @return string[]
      */
     public function selectHeaders(array $accept, string $contentType, bool $isMultipart): array
@@ -53,7 +53,7 @@ class HeaderSelector
         }
 
         if (!$isMultipart) {
-            if($contentType === '') {
+            if ($contentType === '') {
                 $contentType = 'application/json';
             }
 
@@ -96,13 +96,13 @@ class HeaderSelector
     }
 
     /**
-    * Create an Accept header string from the given "Accept" headers array, recalculating all weights
-    *
-    * @param string[] $accept            Array of Accept Headers
-    * @param string[] $headersWithJson   Array of Accept Headers of type "json"
-    *
-    * @return string "Accept" Header (e.g. "application/json, text/html; q=0.9")
-    */
+     * Create an Accept header string from the given "Accept" headers array, recalculating all weights
+     *
+     * @param string[] $accept Array of Accept Headers
+     * @param string[] $headersWithJson Array of Accept Headers of type "json"
+     *
+     * @return string "Accept" Header (e.g. "application/json, text/html; q=0.9")
+     */
     private function getAcceptHeaderWithAdjustedWeight(array $accept, array $headersWithJson): string
     {
         $processedHeaders = [
@@ -112,7 +112,6 @@ class HeaderSelector
         ];
 
         foreach ($accept as $header) {
-
             $headerData = $this->getHeaderAndWeight($header);
 
             if (stripos($headerData['header'], 'application/json') === 0) {
@@ -129,7 +128,7 @@ class HeaderSelector
 
         $hasMoreThan28Headers = count($accept) > 28;
 
-        foreach($processedHeaders as $headers) {
+        foreach ($processedHeaders as $headers) {
             if (count($headers) > 0) {
                 $acceptHeaders[] = $this->adjustWeight($headers, $currentWeight, $hasMoreThan28Headers);
             }
@@ -153,7 +152,7 @@ class HeaderSelector
         if (preg_match('/(.*);\s*q=(1(?:\.0+)?|0\.\d+)$/', $header, $outputArray) === 1) {
             $headerData = [
                 'header' => $outputArray[1],
-                'weight' => (int)($outputArray[2] * 1000),
+                'weight' => (int)((int)$outputArray[2] * 1000),
             ];
         } else {
             $headerData = [
@@ -167,11 +166,11 @@ class HeaderSelector
 
     /**
      * @param array[] $headers
-     * @param float   $currentWeight
-     * @param bool    $hasMoreThan28Headers
+     * @param int $currentWeight
+     * @param bool $hasMoreThan28Headers
      * @return string[] array of adjusted "Accept" headers
      */
-    private function adjustWeight(array $headers, float &$currentWeight, bool $hasMoreThan28Headers): array
+    private function adjustWeight(array $headers, int &$currentWeight, bool $hasMoreThan28Headers): array
     {
         usort($headers, function (array $a, array $b) {
             return $b['weight'] - $a['weight'];
@@ -179,8 +178,7 @@ class HeaderSelector
 
         $acceptHeaders = [];
         foreach ($headers as $index => $header) {
-            if($index > 0 && $headers[$index - 1]['weight'] > $header['weight'])
-            {
+            if ($index > 0 && $headers[$index - 1]['weight'] > $header['weight']) {
                 $currentWeight = $this->getNextWeight($currentWeight, $hasMoreThan28Headers);
             }
 
@@ -196,12 +194,12 @@ class HeaderSelector
 
     /**
      * @param string $header
-     * @param int    $weight
+     * @param int $weight
      * @return string
      */
     private function buildAcceptHeader(string $header, int $weight): string
     {
-        if($weight === 1000) {
+        if ($weight === 1000) {
             return $header;
         }
 
@@ -226,7 +224,7 @@ class HeaderSelector
      * if there is a maximum of 28 "Accept" headers. If we have more than that (which is extremely unlikely), then we fall back to a 1-by-1
      * decrement rule, which will result in quality codes like "q=0.999", "q=0.998" etc.
      *
-     * @param int  $currentWeight varying from 1 to 1000 (will be divided by 1000 to build the quality value)
+     * @param int $currentWeight varying from 1 to 1000 (will be divided by 1000 to build the quality value)
      * @param bool $hasMoreThan28Headers
      * @return int
      */
@@ -240,6 +238,6 @@ class HeaderSelector
             return $currentWeight - 1;
         }
 
-        return $currentWeight - 10 ** floor( log10($currentWeight - 1) );
+        return $currentWeight - 10 ** floor(log10($currentWeight - 1));
     }
 }
